pyproject.toml
[tool.poetry]
name = "anarcpt"
version = "0.1.0"
description = ""
authors = ["\"Nick Muoh\" <\"nickspirit3+github@gmail.com\">"]

[tool.poetry.scripts]
anarcpt = "anarcpt.main:cli"

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
amazon-textract-prettyprinter = "^0.0.10"
awscrt = "^0.12.6"
boto3 = "^1.20.26"
Pillow = "^8.4.0"
loguru = "^0.5.3"
ImageHash = "^4.2.1"
sqlmodel = "^0.0.6"

[tool.poetry.dev-dependencies]
boto3-stubs = {extras = ["s3", "textract"], version = "^1.20.26"}
fs-s3fs = "^1.1.1"
typer = "^0.4.0"
watchdog = "^2.1.6"
alembic = "^1.7.6"

[tool.pyright]
venvPath = "./"
venv = ".venv"
pythonVersion = "3.10"

[tool.mypy]
python_version = "3.10"
python_executable = "./.venv/bin/python"

[[tool.mypy.overrides]]
module = ["Pillow"]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
